+++ packages/server/src/Connection.ts
@@ -119,7 +119,6 @@ export class Connection {
       this.webSocket.close(event?.code, event?.reason)
 
       done()
-
     })
   }