+++ tests/provider/onAuthenticationFailed.js
@@ -22,7 +22,6 @@ context('provider/onAuthenticationFailed', () => {
       name: 'hocuspocus-test',
       document: ydoc,
       WebSocketPolyfill: WebSocket,
-      maxAttempts: 1,
       token: 'SUPER-SECRET-TOKEN',
       onAuthenticationFailed: () => {
         client.destroy()