+++ tests/provider/onMessage.js
@@ -16,7 +16,6 @@ context('provider/onMessage', () => {
       name: 'hocuspocus-test',
       document: ydoc,
       WebSocketPolyfill: WebSocket,
-      maxAttempts: 1,
       onMessage: () => {
         client.destroy()
         server.destroy()
@@ -34,7 +33,6 @@ context('provider/onMessage', () => {
       name: 'hocuspocus-test',
       document: ydoc,
       WebSocketPolyfill: WebSocket,
-      maxAttempts: 1,
     })
 
     client.on('message', () => {