+++ tests/provider/onOpen.js
@@ -22,7 +22,6 @@ context('provider/onOpen', () => {
       name: 'hocuspocus-test',
       document: ydoc,
       WebSocketPolyfill: WebSocket,
-      maxAttempts: 1,
       onOpen: () => {
         client.destroy()
         done()
@@ -36,7 +35,6 @@ context('provider/onOpen', () => {
       name: 'hocuspocus-test',
       document: ydoc,
       WebSocketPolyfill: WebSocket,
-      maxAttempts: 1,
     })
 
     client.on('open', () => {