+++ tests/redis/onCreateDocument.js
@@ -37,7 +37,6 @@ context('redis/onCreateDocument', () => {
       name: 'hocuspocus-test',
       document: ydoc,
       WebSocketPolyfill: WebSocket,
-      maxAttempts: 1,
       broadcast: false,
       // foo.0 = 'bar'
       onSynced: () => {
@@ -58,7 +57,6 @@ context('redis/onCreateDocument', () => {
       name: 'hocuspocus-test',
       document: anotherYdoc,
       WebSocketPolyfill: WebSocket,
-      maxAttempts: 1,
       // foo.0 === 'bar'
       onSynced: () => {
         const value = anotherYdoc.getArray('foo').get(0)