+++ tests/rocksdb/onCreateDocument.js
@@ -34,7 +34,6 @@ context('rocksdb/onCreateDocument', () => {
       name: 'hocuspocus-test',
       document: ydoc,
       WebSocketPolyfill: WebSocket,
-      maxAttempts: 1,
       onSynced: () => {
         const valueBefore = ydoc.getArray('foo').get(0)
         assert.strictEqual(valueBefore, undefined)
@@ -53,7 +52,6 @@ context('rocksdb/onCreateDocument', () => {
       name: 'hocuspocus-test',
       document: anotherYdoc,
       WebSocketPolyfill: WebSocket,
-      maxAttempts: 1,
       onSynced: () => {
         const value = anotherYdoc.getArray('foo').get(0)
         assert.strictEqual(value, 'bar')