+++ packages/logger/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-alpha.49](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/extension-logger@1.0.0-alpha.48...@hocuspocus/extension-logger@1.0.0-alpha.49) (2021-10-15)
+
+**Note:** Version bump only for package @hocuspocus/extension-logger
+
+
+
+
+
+# [1.0.0-alpha.48](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/extension-logger@1.0.0-alpha.47...@hocuspocus/extension-logger@1.0.0-alpha.48) (2021-10-08)
+
+**Note:** Version bump only for package @hocuspocus/extension-logger
+
+
+
+
+
+# [1.0.0-alpha.47](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/extension-logger@1.0.0-alpha.46...@hocuspocus/extension-logger@1.0.0-alpha.47) (2021-09-23)
+
+**Note:** Version bump only for package @hocuspocus/extension-logger
+
+
+
+
+
+# [1.0.0-alpha.46](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/extension-logger@1.0.0-alpha.45...@hocuspocus/extension-logger@1.0.0-alpha.46) (2021-09-23)
+
+**Note:** Version bump only for package @hocuspocus/extension-logger
+
+
+
+
+
 # [1.0.0-alpha.45](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/extension-logger@1.0.0-alpha.44...@hocuspocus/extension-logger@1.0.0-alpha.45) (2021-09-17)
 
 **Note:** Version bump only for package @hocuspocus/extension-logger